Remove WalletInitializer
